+++ R/time_cut.R
@@ -50,11 +50,11 @@ cut_time_using_ops <- function(x, width, from = NULL, to = NULL){
     from <- collapse::fmin(x, na.rm = TRUE)
   } else {
     from <- time_cast(from, x)
-    x[x < from] <- NA
+    x[cheapr::val_find(x < from, TRUE)] <- NA
   }
   if (!to_missing){
     to <- time_cast(to, x)
-    x[x >= to] <- NA
+    x[cheapr::val_find(x >= to, TRUE)] <- NA
   }
   num <- timespan_num(width)
   units <- timespan_unit(width)
@@ -165,29 +165,35 @@ cut_time_using_ops <- function(x, width, from = NULL, to = NULL){
 #' Cut dates and datetimes into regularly spaced date or datetime intervals
 #'
 #' @description
-#' Useful functions especially for when plotting time-series.
-#' `time_cut` makes approximately `n` groups of equal time range.
+#' Useful functions especially for when plotting time-series. \cr
+#'
+#' `time_cut_n` makes approximately `n` groups of equal time range.
 #' It prioritises the highest time unit possible, making axes look
-#' less cluttered and thus prettier.
-#' `time_breaks` returns only the breaks.
+#' less cluttered and thus prettier. \cr
+#'
+#' `time_breaks` returns only the breakpoints.
+#'
+#' `time_breakpoints` is a newer and faster alternative to
+#' `time_breaks` which differs in
+#' that it calls `range()` on the input data and therefore need only work with
+#' a vector of 2 values, unlike `time_breaks` which requires more data points
+#' to create better looking breaks.
 #'
 #' @details
 #' To retrieve regular time breaks that simply spans the range of `x`,
-#' use `time_seq()` or `time_aggregate()`.
-#' This can also be achieved in `time_cut()` by supplying `n = Inf`.
+#' use `time_seq()`to manually specify the range and time width or
+#' `time_grid()` to use the range of the supplied data.
 #'
-#' By default `time_cut()` will try to find
-#'  the prettiest way of cutting the interval by
+#'
+#' By default `time_cut_n()` will try to find
+#' the 'prettiest' way of cutting the interval by
 #' trying to cut the date/date-times into
 #' groups of the highest possible time units,
 #' starting at years and ending at milliseconds.
 #'
-#' When `x` is a numeric vector, `time_cut` will behave similar to `time_cut`
-#' except for 3 things:
-#' * The intervals are all right-open and of equal width.
-#' * The left value of the leftmost interval is always `min(x)`.
-#' * Up to `n` breaks are created, i.e `<= n` breaks. This is to prioritise
-#'   pretty breaks.
+#' `time_breakpoints` does the same but using a different internal method.
+#'
+#'
 #' @inheritParams time_grid
 #' @param n Number of breaks.
 #' @param time_floor Logical. Should the initial date/datetime be
@@ -197,8 +203,8 @@ cut_time_using_ops <- function(x, width, from = NULL, to = NULL){
 #' This is only used when `time_floor = TRUE`.
 #'
 #' @returns
-#' `time_breaks` returns a vector of breaks. \cr
-#' `time_cut` returns either a vector or `time_interval`. \cr
+#' `time_breaks` and `time_breakpoints` both return a vector of breakpoints \cr
+#' `time_cut_n` and `time_cut_width` returns a `time_interval`
 #'
 #' @examples
 #' library(timeplyr)
@@ -211,21 +217,21 @@ cut_time_using_ops <- function(x, width, from = NULL, to = NULL){
 #' time_cut_n(1:10, n = 5)
 #'
 #' # Easily create custom time breaks
-#' df <- nycflights13::flights %>%
-#'   f_slice_sample(n = 100) %>%
-#'   with_local_seed(.seed = 8192821) %>%
-#'   f_select(time_hour) %>%
-#'   fastplyr::f_arrange(time_hour) %>%
+#' df <- nycflights13::flights |>
+#'   f_slice_sample(n = 100) |>
+#'   with_local_seed(.seed = 8192821) |>
+#'   f_select(time_hour) |>
+#'   fastplyr::f_arrange(time_hour) |>
 #'   mutate(date = as_date(time_hour))
 #'
 #' # time_cut_n() and time_breaks() automatically find a
 #' # suitable way to cut the data
-#' time_cut_n(df$date) %>%
+#' time_cut_n(df$date) |>
 #'   interval_count()
 #' # Works with datetimes as well
-#' time_cut_n(df$time_hour, n = 5) %>%
+#' time_cut_n(df$time_hour, n = 5) |>
 #'   interval_count()
-#' time_cut_n(df$date, timespan = "month") %>%
+#' time_cut_n(df$date, timespan = "month") |>
 #'   interval_count()
 #' # Just the breaks
 #' time_breaks(df$date, n = 5, timespan = "month")
@@ -233,35 +239,20 @@ cut_time_using_ops <- function(x, width, from = NULL, to = NULL){
 #' cut_dates <- time_cut_n(df$date)
 #' date_breaks <- time_breaks(df$date)
 #'
-#' # When n = Inf it should be equivalent to using time_cut_width
-#' identical(time_cut_n(df$date, n = Inf, "month"),
-#'           time_cut_width(df$date, "month"))
 #' # To get exact breaks at regular intervals, use time_grid
 #' weekly_breaks <- time_grid(
 #'   df$date, "5 weeks",
 #'   from = floor_date(min(df$date), "week", week_start = 1)
 #' )
 #' weekly_labels <- format(weekly_breaks, "%b-%d")
-#' df %>%
-#'   time_by(date, "week", .name = "date") %>%
-#'   f_count() %>%
-#'   mutate(date = interval_start(date)) %>%
+#' df |>
+#'   time_by(date, "week", .name = "date") |>
+#'   f_count() |>
+#'   mutate(date = interval_start(date)) |>
 #'   ggplot(aes(x = date, y = n)) +
 #'   geom_bar(stat = "identity") +
 #'   scale_x_date(breaks = weekly_breaks,
 #'                labels = weekly_labels)
-#' @rdname time_cut
-#' @export
-time_cut <- function(x, n = 5, timespan = NULL,
-                     from = NULL, to = NULL,
-                     time_floor = FALSE,
-                     week_start = getOption("lubridate.week.start", 1)){
-  lifecycle::deprecate_soft("1.0.0", "time_cut()", "time_cut_n()")
-  time_cut_n(x, n = n, timespan = timespan,
-             from = from, to = to,
-             time_floor = time_floor,
-             week_start = week_start)
-}
 #' @rdname time_cut
 #' @export
 time_cut_n <- function(x, n = 5, timespan = NULL,
@@ -322,4 +313,77 @@ time_breaks <- function(x, n = 5, timespan = NULL,
                       time_floor = time_floor,
                       week_start = week_start)
   breaks[["breaks"]]
+}
+
+#' @rdname time_cut
+#' @export
+time_breakpoints <- function(x, n = 10){
+
+  rng <- as.double(collapse::frange(x))
+  time_rng <- time_cast(rng, x)
+
+  # get_breaks() returns pretty numeric breaks efficiently
+  # we can use this as a good starting point
+
+  breakpoints <- cheapr::get_breaks(rng, n = n, pretty = TRUE)
+  guess <- time_cast(breakpoints, x)
+
+  # This should never be the case but include it just to be safe
+  if (length(guess) < 2){
+    return(guess)
+  }
+
+  # This can happen after coercion from e.g. vector of double breakpoints to
+  # integer-based year-quarters
+  if (guess[1] == guess[2]){
+    return(unique(guess))
+  }
+
+  left <- time_rng[1]
+  right <- time_rng[2]
+
+  if (is_time(guess)){
+
+    # In order of decreasing unit (years, months, etc...)
+    # Stop when the correct (approximate)
+    # time granularity of breakpoint widths is found
+    # granularity() can't be used as it will return a lower unit if the
+    # granularity at a higher unit isn't exact
+
+    for (unit in rev(time_units)){
+      span <- new_timespan(unit, 1)
+      width <- time_diff(guess[1], guess[2], span)
+      if (width > 1){
+        break
+      }
+    }
+
+    # If width is sub-fractional seconds
+    if (unit == "seconds" && width < 1){
+      width <- span * floor_nearest_n(width, 10^-(ceiling(abs(log10(width)))))
+    } else {
+      width <- span * ceiling(width)
+      left <- time_floor(left, span)
+      right <- time_ceiling(right, span)
+    }
+  } else {
+    width <- time_diff(guess[1], guess[2])
+    left <- time_floor(left, width)
+    right <- time_ceiling(right, width)
+  }
+
+  time_seq_v(left, right, width)
+}
+
+#' @rdname time_cut
+#' @export
+time_cut <- function(x, n = 5, timespan = NULL,
+                     from = NULL, to = NULL,
+                     time_floor = FALSE,
+                     week_start = getOption("lubridate.week.start", 1)){
+  lifecycle::deprecate_warn("1.0.0", "time_cut()", "time_cut_n()")
+  time_cut_n(x, n = n, timespan = timespan,
+             from = from, to = to,
+             time_floor = time_floor,
+             week_start = week_start)
 }