+++ R/time_ggplot.R
@@ -30,29 +30,29 @@
 #' collapse::set_collapse(nthreads = 1L)
 #' }
 #' # It's as easy as this
-#' AirPassengers %>%
-#'   ts_as_tbl() %>%
+#' AirPassengers |>
+#'   ts_as_tbl() |>
 #'   time_ggplot(time, value)
 #'
 #' # And this
-#' EuStockMarkets %>%
-#'   ts_as_tbl() %>%
+#' EuStockMarkets |>
+#'   ts_as_tbl() |>
 #'   time_ggplot(time, value, group)
 #'
 #' # Converting this to monthly averages
 #'
-#' EuStockMarkets %>%
-#'   ts_as_tbl() %>%
-#'   mutate(month = year_month_decimal(time)) %>%
+#' EuStockMarkets |>
+#'   ts_as_tbl() |>
+#'   mutate(month = year_month_decimal(time)) |>
 #'   summarise(avg = mean(value),
-#'             .by = c(group, month)) %>%
+#'             .by = c(group, month)) |>
 #'   time_ggplot(month, avg, group)
 #'
 #' # zoo example
 #' x.Date <- as.Date("2003-02-01") + c(1, 3, 7, 9, 14) - 1
 #' x <- zoo::zoo(rnorm(5), x.Date)
-#' x %>%
-#'   ts_as_tbl() %>%
+#' x |>
+#'   ts_as_tbl() |>
 #'   time_ggplot(time, value)
 #' \dontshow{
 #' data.table::setDTthreads(threads = .n_dt_threads)
@@ -61,44 +61,49 @@
 #' @export
 time_ggplot <- function(data, time, value, group = NULL,
                         facet = FALSE, geom = ggplot2::geom_line, ...){
+
+  if (!is.logical(facet)){
+   cli::cli_abort("{.arg facet} must be a logical scalar")
+  }
+
   # Tidyselect variables
-  time <- tidy_select_pos(data, !!enquo(time))
-  value <- tidy_select_pos(data, !!enquo(value))
-  group <- tidy_select_pos(data, !!enquo(group))
+  time <- tidy_select_names(data, !!enquo(time))
+  value <- tidy_select_names(data, !!enquo(value))
+  group <- tidy_select_names(data, !!enquo(group))
   time <- names(time)
   value <- names(value)
   group <- names(group)
   time_var <- data[[time]]
   # Pretty x-axis breaks
-  time_breaks <- time_breaks(time_var, n = 7, time_floor = TRUE)
+  breakpoints <- time_breakpoints(time_var, n = 10)
   if (is_datetime(time_var)){
-    x_scale <- ggplot2::scale_x_datetime(breaks = time_breaks,
+    x_scale <- ggplot2::scale_x_datetime(breaks = breakpoints,
                                          labels = scales::label_date_short())
   } else if (is_date(time_var)){
-    x_scale <- ggplot2::scale_x_date(breaks = time_breaks,
+    x_scale <- ggplot2::scale_x_date(breaks = breakpoints,
                                      labels = scales::label_date_short())
   } else if (is_year_month(time_var)){
-    x_scale <- scale_x_year_month(breaks = function(x) cheapr::get_breaks(as.double(x)))
+    x_scale <- scale_x_year_month(breaks = breakpoints)
   } else if (is_year_quarter(time_var)){
-    x_scale <- scale_x_year_quarter(breaks = function(x) cheapr::get_breaks(as.double(x)))
+    x_scale <- scale_x_year_quarter(breaks = breakpoints)
   } else {
-    x_scale <- ggplot2::scale_x_continuous(breaks = function(x) cheapr::get_breaks(as.double(x)))
+    x_scale <- ggplot2::scale_x_continuous(breaks = breakpoints)
   }
   # Concatenate group names together
   if (length(group) > 1L){
     group_nm <- unique_col_name(data, "group")
     group_col <- list(
       df_paste_names(
-        fastplyr::f_select(df_ungroup(data), .cols = group)
+        fastplyr::f_select(fastplyr::f_ungroup(data), .cols = group)
       )
     )
     names(group_col) <- group_nm
-    data <- df_add_cols(data, cols = group_col)
+    data <- cheapr::df_modify(data, cols = group_col)
   } else {
     group_nm <- group
   }
   # Time-series plot
-  out <- data %>%
+  out <- data |>
     ggplot2::ggplot(ggplot2::aes(x = .data[[time]],
                                  y = .data[[value]])) +
     ggplot2::theme_minimal() +
@@ -107,11 +112,10 @@ time_ggplot <- function(data, time, value, group = NULL,
   if (length(group) > 0L){
     if (facet){
       # Add a new col every 6 rows
-      facet_ncol <- (collapse::fnunique(fpluck(data, group_nm)) %/% 6) + 1
+      facet_ncol <- (collapse::fnunique(data[[group_nm]]) %/% 6) + 1
       out <- out +
         do.call(geom, extra_gg_args) +
-        ggplot2::facet_wrap(group, ncol = facet_ncol,
-                            scales = "free_y")
+        ggplot2::facet_wrap(group, ncol = facet_ncol, scales = "free_y")
     } else {
       out <- out +
         do.call(geom, c(list(ggplot2::aes(col = .data[[group_nm]])), extra_gg_args))