+++ R/time_seq_id.R
@@ -68,10 +68,11 @@
 time_seq_id <- function(x, timespan = granularity(x),
                         threshold = 1, g = NULL, na_skip = TRUE,
                         rolling = TRUE, switch_on_boundary = FALSE){
+
   check_is_time_or_num(x)
+
   g <- GRP2(g)
-  timespan <- timespan(timespan)
-  time_num <- timespan_num(timespan)
+
   if (is_whole_number(threshold)){
     threshold <- as.integer(threshold)
   }
@@ -82,7 +83,7 @@ time_seq_id <- function(x, timespan = granularity(x),
     fill <- -Inf
   }
   # Elapsed time
-  telapsed <- time_elapsed(x, timespan, g = g,
+  telapsed <- time_elapsed(x, timespan(timespan), g = g,
                            rolling = rolling, na_skip = na_skip,
                            fill = fill)
   if (rolling){
@@ -92,16 +93,17 @@ time_seq_id <- function(x, timespan = granularity(x),
       over_threshold <- cppdoubles::double_gt(telapsed, threshold)
     }
   } else {
-    dt <- new_dt(x = telapsed, group_id = fastplyr::group_id(g), .copy = FALSE)
-    group_id_col <- names(dt)[names(dt) == "group_id"]
-    dt[, ("over") :=
-         lapply(
-           .SD, function(x)
-             cpp_roll_time_threshold(x, threshold = threshold, switch_on_boundary = switch_on_boundary)
-         ),
-       by = group_id_col,
-       .SDcols = "x"]
-    over_threshold <- dt[["over"]]
+    if (is.null(g)){
+      over_threshold <- cpp_roll_time_threshold(
+        telapsed, threshold = threshold, switch_on_boundary = switch_on_boundary
+      )
+    } else {
+      over_threshold <- fastplyr::new_tbl(x = telapsed, group_id = GRP_group_id(g)) |>
+        fastplyr::f_mutate(.over = cpp_roll_time_threshold(
+          x, threshold = threshold, switch_on_boundary = switch_on_boundary
+        ), .by = dplyr::any_of("group_id"), .order = FALSE) |>
+        fastplyr::f_pull(.cols = ".over")
+    }
   }
   collapse::fcumsum(over_threshold, g = g, na.rm = na_skip) + 1L
 }