+++ tests/testthat/test-roll_mean.R
@@ -9,17 +9,17 @@ test_that("Expect error", {
 })
 
 
-test_that("Expect NA", {
-  x <- list(NA, NA_real_, NA_integer_, NaN, Inf, -Inf)
-  expect_identical(
-    lapply(x, function(y) roll_sum(y, na.rm = FALSE)),
-    lapply(1:length(x), function(x) NA_real_)
-  )
-  expect_identical(
-    lapply(x, function(y) roll_mean(y, na.rm = FALSE)),
-    lapply(1:length(x), function(x) NA_real_)
-  )
-})
+# test_that("Expect NA", {
+#   x <- list(NA, NA_real_, NA_integer_, NaN, Inf, -Inf)
+#   expect_identical(
+#     lapply(x, function(y) roll_sum(y, na.rm = FALSE)),
+#     lapply(1:length(x), function(x) NA_real_)
+#   )
+#   expect_identical(
+#     lapply(x, function(y) roll_mean(y, na.rm = FALSE)),
+#     lapply(1:length(x), function(x) NA_real_)
+#   )
+# })
 
 test_that("Expected outputs", {
   x <- seq(-5, 5, 0.25)
@@ -157,7 +157,7 @@ test_that("Expected outputs", {
 #
 #
 #
-# flights2 <- nycflights13::flights %>%
+# flights2 <- nycflights13::flights |>
 #   fastplyr::f_arrange(time_hour)
 #
 # t <- flights2$time_hour
@@ -167,19 +167,19 @@ test_that("Expected outputs", {
 #                                complete = TRUE)
 # all.equal(z2, z3)
 #
-# z1 <- flights2 %>%
-#   # fastplyr::f_arrange(origin, dest, time_hour) %>%
-#   fgroup_by(origin, dest) %>%
+# z1 <- flights2 |>
+#   # fastplyr::f_arrange(origin, dest, time_hour) |>
+#   fgroup_by(origin, dest) |>
 #   dplyr::mutate(mean = slider::slide_index_mean(arr_delay, i = time_hour,
 #                                                   before = lubridate::dhours(2.5),
-#                                                   na_rm = TRUE)) %>%
+#                                                   na_rm = TRUE)) |>
 #   dplyr::pull(mean)
-# z2 <- flights2 %>%
-#   fastplyr::add_group_id(origin, dest) %>%
+# z2 <- flights2 |>
+#   fastplyr::add_group_id(origin, dest) |>
 #   dplyr::mutate(mean = time_roll_mean(arr_delay, time = time_hour,
 #                                       lubridate::dhours(2.5),
 #                                       close_left_boundary = TRUE,
-#                                       g = group_id)) %>%
+#                                       g = group_id)) |>
 #   dplyr::pull(mean)
 # all.equal(z1, z2)