+++ NEWS.md
@@ -1,3 +1,14 @@
+# timeplyr 1.1.1
+
+- Rolling calculations have been fixed to work with the new 
+rolling data.table functions.
+
+- `time_add` gains the `n` arg to help with adding timespans to 
+time-based vectors more easily.
+
+- New function `time_breakpoints` as a simpler alternative to `time_breaks` that
+works much better with `ggplot2`
+
 # timeplyr 1.1.0
 
 - New options 'xfirst' and 'xlast' for 'timeplyr.roll_month' to signify